requirements.txt
discord.py
twitchio
redis
asyncpg
psycopg2-binary
python-dotenv
